requirements.txt
streamlit
anthropic
pytesseract
Pillow
python-dotenv
streamlit-js-eval
